chore: extract createCommonSection
Move common section of rockcraft.yaml into the base class
